+++ crates/biome_cli/src/commands/mod.rs
@@ -1,14 +1,17 @@
 use crate::cli_options::{cli_options, CliOptions, ColorsArg};
+use crate::diagnostics::DeprecatedConfigurationFile;
 use crate::logging::LoggingKind;
-use crate::{LoggingLevel, VERSION};
+use crate::{CliDiagnostic, LoggingLevel, VERSION};
+use biome_console::{markup, Console, ConsoleExt};
+use biome_diagnostics::PrintDiagnostic;
 use biome_service::configuration::json::JsonFormatter;
 use biome_service::configuration::vcs::VcsConfiguration;
 use biome_service::configuration::{
     configuration, files_configuration, formatter_configuration, javascript::javascript_formatter,
     json::json_formatter, linter_configuration, vcs::vcs_configuration, FilesConfiguration,
-    FormatterConfiguration, JavascriptFormatter, LinterConfiguration,
+    FormatterConfiguration, JavascriptFormatter, LinterConfiguration, LoadedConfiguration,
 };
-use biome_service::Configuration;
+use biome_service::{Configuration, ConfigurationDiagnostic, WorkspaceError};
 use bpaf::Bpaf;
 use std::ffi::OsString;
 use std::path::PathBuf;
@@ -90,6 +93,17 @@ pub enum BiomeCommand {
         /// Example: `echo 'let a;' | biome check --stdin-file-path=file.js`
         #[bpaf(long("stdin-file-path"), argument("PATH"), hide_usage)]
         stdin_file_path: Option<String>,
+
+        /// When set to true, only the files that have been changed compared to your `defaultBranch`
+        /// configuration will be linted.
+        #[bpaf(long("changed"), switch)]
+        changed: bool,
+
+        /// Use this to specify the base branch to compare against when you're using the --changed
+        /// flag and the `defaultBranch` is not set in your biome.json
+        #[bpaf(long("since"), argument("REF"))]
+        since: Option<String>,
+
         /// Single file, single path or list of paths
         #[bpaf(positional("PATH"), many)]
         paths: Vec<OsString>,
@@ -121,6 +135,14 @@ pub enum BiomeCommand {
         /// Example: `echo 'let a;' | biome lint --stdin-file-path=file.js`
         #[bpaf(long("stdin-file-path"), argument("PATH"), hide_usage)]
         stdin_file_path: Option<String>,
+        /// When set to true, only the files that have been changed compared to your `defaultBranch`
+        /// configuration will be linted.
+        #[bpaf(long("changed"), switch)]
+        changed: bool,
+        /// Use this to specify the base branch to compare against when you're using the --changed
+        /// flag and the `defaultBranch` is not set in your biome.json
+        #[bpaf(long("since"), argument("REF"))]
+        since: Option<String>,
         /// Single file, single path or list of paths
         #[bpaf(positional("PATH"), many)]
         paths: Vec<OsString>,
@@ -157,6 +179,16 @@ pub enum BiomeCommand {
         #[bpaf(switch)]
         write: bool,
 
+        /// When set to true, only the files that have been changed compared to your `defaultBranch`
+        /// configuration will be linted.
+        #[bpaf(long("changed"), switch)]
+        changed: bool,
+
+        /// Use this to specify the base branch to compare against when you're using the --changed
+        /// flag and the `defaultBranch` is not set in your biome.json
+        #[bpaf(long("since"), argument("REF"))]
+        since: Option<String>,
+
         /// Single file, single path or list of paths.
         #[bpaf(positional("PATH"), many)]
         paths: Vec<OsString>,
@@ -180,6 +212,16 @@ pub enum BiomeCommand {
         configuration: Configuration,
         #[bpaf(external, hide_usage)]
         cli_options: CliOptions,
+
+        /// When set to true, only the files that have been changed compared to your `defaultBranch`
+        /// configuration will be linted.
+        #[bpaf(long("changed"), switch)]
+        changed: bool,
+
+        /// Use this to specify the base branch to compare against when you're using the --changed
+        /// flag and the `defaultBranch` is not set in your biome.json
+        #[bpaf(long("since"), argument("REF"))]
+        since: Option<String>,
 
         /// Single file, single path or list of paths
         #[bpaf(positional("PATH"), many)]
@@ -310,4 +352,42 @@ impl BiomeCommand {
             | BiomeCommand::PrintSocket => LoggingKind::default(),
         }
     }
+}
+
+/// It accepts a [LoadedConfiguration] and it prints the diagnostics emitted during parsing and deserialization.
+///
+/// If it contains errors, it return an error.
+pub(crate) fn validate_configuration_diagnostics(
+    loaded_configuration: &LoadedConfiguration,
+    console: &mut dyn Console,
+    verbose: bool,
+) -> Result<(), CliDiagnostic> {
+    if let Some(file_path) = loaded_configuration
+        .file_path
+        .as_ref()
+        .and_then(|f| f.file_name())
+        .and_then(|f| f.to_str())
+    {
+        if file_path == "rome.json" {
+            let diagnostic = DeprecatedConfigurationFile::new(file_path);
+            console.error(markup!{
+                {if verbose { PrintDiagnostic::verbose(&diagnostic) } else { PrintDiagnostic::simple(&diagnostic) }}
+            });
+        }
+    }
+    let diagnostics = loaded_configuration.as_diagnostics_iter();
+    for diagnostic in diagnostics {
+        console.error(markup!{
+            {if verbose { PrintDiagnostic::verbose(diagnostic) } else { PrintDiagnostic::simple(diagnostic) }}
+        });
+    }
+    if loaded_configuration.has_errors() {
+        return Err(CliDiagnostic::workspace_error(
+            WorkspaceError::Configuration(ConfigurationDiagnostic::invalid_configuration(
+                "Biome exited because the configuration resulted in errors. Please fix them.",
+            )),
+        ));
+    }
+
+    Ok(())
 }