+++ crates/biome_js_analyze/tests/specs/suspicious/noEmptyInterface/valid.ts
@@ -9,4 +9,8 @@ interface Baz extends Foo, Bar {}
 // See https://github.com/biomejs/biome/issues/959
 declare module "external" {
   export interface App extends Services {}
+
+  global {
+    export interface App extends Services {}
+  }
 }